pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ManagerX"
version = "2026.0.1"
description = "A powerful Discord bot for server management and fun."
readme = "README.md"
requires-python = ">=3.8"
license = {text = "GPL-3.0-or-later"}
authors = [
    {name = "OPPRO.NET Development", email = "development@oppro-network.de"},
    {name = "OPPRO.NET Network", email = "contact@oppro-network.de"}
]
classifiers = [
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Operating System :: OS Independent"
]

# Korrigierte Syntax: Direktes Array
dependencies = [
    "ezcord==0.7.4",
    "py-cord==2.7.0",
    "aiosqlite==0.22.1",
    "aiohttp==3.13.2",
    "aiocache==0.12.3",
    "propcache==0.4.1",
    "requests==2.32.5",
    "wikipedia==1.4.0",
    "beautifulsoup4==4.14.3",
    "soupsieve==2.8.1",
    "yarl==1.22.0",
    "frozenlist==1.8.0",
    "h11==0.16.0",
    "multidict==6.7.0",
    "SimpleColoredLogs==1.3.0",
    "FastAPI==0.100.0",
    "uvicorn==0.24.0",
    "Jinja2==3.1.6",
    "MarkupSafe==3.0.3",
    "starlette==0.50.0",
    "timedelta==2020.12.3",
    "ManagerX-Handler==1.0.4",
    "ManagerX-DevTools==1.0.0"
]

[project.urls]
Homepage = "https://github.com/Oppro-net-Development/ManagerX"
Documentation = "https://docs.oppro-network.de/"
Changelog = "https://docs.oppro-network.de/en/latest/releases/index.html"
BugTracker = "https://github.com/Oppro-net-Development/ManagerX/issues"

[project.optional-dependencies]
dev = [
    "python-dotenv==1.2.1",
    "click==8.3.1",
    "colorama==0.4.6",
    "typing_extensions==4.15.0",
    "typing-inspection==0.4.2",
    "attrs==25.4.0",
    "annotated-types==0.7.0",
    "anyio==4.12.0",
    "certifi==2025.11.12",
    "charset-normalizer==3.4.4",
    "idna==3.11",
    "urllib3==2.6.2",
    "build==1.3.0",
    "twine==4.2.0"
]

docs = [
    "sphinx<8,>=6",
    "pydata-sphinx-theme",
    "sphinx-autodoc-typehints",
    "myst-parser<3,>=2",
    "sphinx-copybutton"
]

all = [
    "ezcord==0.7.4",
    "py-cord==2.7.0",
    "aiosqlite==0.22.1",
    "aiohttp==3.13.2",
    "aiocache==0.12.3",
    "propcache==0.4.1",
    "requests==2.32.5",
    "wikipedia==1.4.0",
    "beautifulsoup4==4.14.3",
    "soupsieve==2.8.1",
    "yarl==1.22.0",
    "frozenlist==1.8.0",
    "h11==0.16.0",
    "multidict==6.7.0",
    "python-dotenv==1.2.1",
    "click==8.3.1",
    "colorama==0.4.6",
    "typing_extensions==4.15.0",
    "typing-inspection==0.4.2",
    "attrs==25.4.0",
    "annotated-types==0.7.0",
    "anyio==4.12.0",
    "certifi==2025.11.12",
    "charset-normalizer==3.4.4",
    "idna==3.11",
    "urllib3==2.6.2",
    "Jinja2==3.1.6",
    "MarkupSafe==3.0.3",
    "starlette==0.50.0",
    "FastAPI==0.100.0",
    "uvicorn==0.24.0",
    "SimpleColoredLogs==1.3.0",
    "timedelta==2020.12.3",
    "sphinx<8,>=6",
    "pydata-sphinx-theme",
    "sphinx-autodoc-typehints",
    "myst-parser<3,>=2",
    "sphinx-copybutton",
    "build==1.3.0",
    "twine==4.2.0"
]

[project.scripts]
# Standard-Weg für CLI-Befehle
managerx-cli = "managerx.cli:main"
managerx = "managerx.__main__:main"

[tool.setuptools]
# Definiert, wo der Quellcode liegt
package-dir = {"src" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["managerx*", "plugins*", "DevTools*", "cogs*"]
exclude = ["tests*"]